DT-400 Split allocate POM tests into API test / integration test
